[DDW-821] Moves remaining wallet-helpers into Wallets domain
